chore: reduce the size of Init Process
