+++ pkg/testworkflows/testworkflowprocessor/action/actiontypes/lite/utils.go
@@ -1,36 +1,5 @@
 package lite
 
-import (
-	"fmt"
-
-	corev1 "k8s.io/api/core/v1"
-)
-
-func EnvName(group string, computed bool, sensitive bool, name string) string {
-	suffix := ""
-	if computed {
-		suffix = "C"
-	}
-	if sensitive {
-		suffix += "S"
-	}
-	return fmt.Sprintf("_%s%s_%s", group, suffix, name)
-}
-
-func EnvVar(group string, computed, sensitive bool, name, value string) corev1.EnvVar {
-	return corev1.EnvVar{
-		Name:  EnvName(group, computed, sensitive, name),
-		Value: value,
-	}
-}
-
-func EnvVarFrom(group string, computed, sensitive bool, name string, value corev1.EnvVarSource) corev1.EnvVar {
-	return corev1.EnvVar{
-		Name:      EnvName(group, computed, sensitive, name),
-		ValueFrom: &value,
-	}
-}
-
 type LiteActionList []LiteAction
 
 func NewLiteActionList() LiteActionList {
